requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
gspread==5.12.0
google-auth==2.29.0
pytz==2024.1